github workflows: simplify test matrix
This makes it so that instead of running the full CI build
on each possible OS, we only run our actual tests. The full
CI build is run on a different (usually subset) list of OS'.

The reason for this seperation is that we may want to add
many supported operating systems or many go versions
for which we should run tests. This should speed up builds
(and save money) without losing much information.